+++ src/resolvers/mutation/index.js
@@ -3,6 +3,7 @@ import closetMutation from './closet';
 import outfitMutation from './outfit';
 import paymentMutation from './payment';
 import requestMutation from './request';
+import stylistMutation from './stylist';
 import subscriptionMutation from './subscription';
 import userMutation from './user';
 // all the mutation
@@ -14,6 +15,7 @@ const mutation = {
   ...requestMutation,
   ...closetMutation,
   ...outfitMutation,
+  ...stylistMutation,
 };
 
 export default mutation;